+++ crates/stackable-operator/src/lib.rs
@@ -27,7 +27,6 @@ pub mod logging;
 pub mod memory;
 pub mod namespace;
 pub mod pod_utils;
-pub mod product_config_utils;
 pub mod product_logging;
 pub mod role_utils;
 pub mod status;